tests/wpt/tests/html/canvas/element/path-objects/2d.path.arc.twopie.6.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.path.arc.twopie.6</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.arc.twopie.6</h1>
<p class="desc">arc() draws correctly when start = 5 and end = start + 2pi+e and clockwise</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("arc() draws correctly when start = 5 and end = start + 2pi+e and clockwise");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = '#fff';
  ctx.fillRect(0, 0, 100, 50);
  ctx.fillStyle = '#000';
  ctx.beginPath();
  ctx.moveTo(50, 25);
  ctx.arc(50, 25, 50, 5, 5 + 2*Math.PI, false);
  ctx.closePath();
  ctx.fill();
  _assertPixel(canvas, 5,25, 0,0,0,255);

});
</script>
